Idk how to name this commit

+++ backend/main.py
@@ -18,7 +18,7 @@ from typing import Optional, List
 from utils.crypt import Encoder, Decoder
 from utils.helpers import Parser, Reader
 from sockmanager import RoomsManager, Room
-from models import RoomReq, RoomAuth, MsgKeysCreate, MsgKeysGet
+from models import RoomReq, RoomAuth, MsgKeysCreate, MsgKeysGet, UserBlock
 from database import Database
 from secrets import token_hex
 
@@ -49,16 +49,16 @@ async def home():
 @app.post("/")
 async def create_room(room: RoomReq):
     try:
-        plainName, plainPassword = parser.parse_room_data(room)
-        hashed_name = encoder.hash_name(plainName)
+        plainName, plainPassword = await parser.parse_room_data(room)
+        hashed_name = await encoder.hash_name(plainName)
         link = encoder.gen_hash_link(hashed_name)
-        database.create_room(
+        await database.create_room(
             {
                 'name': hashed_name,
                 'password': encoder.hash_password(plainPassword)
             }
         )
-        sessionCookie = encoder.encode_session(
+        sessionCookie = await encoder.encode_session(
             hashed_name, plainPassword, 'Admin', encoder.key
         )
         return JSONResponse(
@@ -80,39 +80,43 @@ async def create_room(room: RoomReq):
 
 
 @app.get('/rooms/{name}/auth')
-async def room_auth(name: str, session: Optional[str] = Cookie(None)):
+async def room_session_auth(name: str, session: Optional[str] = Cookie(None)):
     try:
         if session:
-            hashed_name = parser.parse_link_hash(name)
-            room = database.get_room_by_name(hashed_name)
-            if decoder.verify_session(hashed_name, room.password, session):
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
+            user = await database.get_user_by_name(decoder.decode_session(session)['username'])
+            if decoder.verify_session(hashed_name, room.password, session) and user.status:
                 return JSONResponse(
                     status_code=status.HTTP_302_FOUND,
-                    content={"auth": True},
+                    content={"Auth": True},
                     headers={
                         'Content-Type': 'application/json',
                         'Location': f'/rooms/{name}',
                     }
                 )
-        return HTMLResponse(
-            templates['room_auth'],
-            status_code=status.HTTP_401_UNAUTHORIZED
+        return JSONResponse(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            content={"Unauthorized": "Session doesn't exist"},
+            headers={
+                'Content-Type': 'application/json',
+            }
         )
-    except Exception as error:
+    except ValueError as error:
         raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            status_code=status.HTTP_400_BAD_REQUEST,
             detail={
-                "Code": 500,
-                "error": error
+                "Code": 400,
+                "Error": error
             })
 
 
 @app.post('/rooms/{name}/auth')
-async def validate_room_password(name: str, auth: RoomAuth):
+async def room_password_auth(name: str, auth: RoomAuth):
     try:
-        hashed_name = parser.parse_link_hash(name)
-        json_data = jsonable_encoder(auth)
-        room = database.get_room_by_name(hashed_name)
+        hashed_name = await parser.parse_link_hash(name)
+        json_data = await jsonable_encoder(auth)
+        room = await database.get_room_by_name(hashed_name)
         password, username = json_data['password'], json_data['username']
         if decoder.verify_password(password, room.password):
             if not database.check_user(username, room.id):
@@ -121,12 +125,12 @@ async def validate_room_password(name: str, auth: RoomAuth):
                     "admin": False,
                     "room_id": room.id,
                 }, room.id)
-                sessionCookie = encoder.encode_session(
+                sessionCookie = await encoder.encode_session(
                     hashed_name, password, username, ''
                 )
                 return JSONResponse(
                     status_code=status.HTTP_302_FOUND,
-                    content={'Auth': True},
+                    content={'User': username},
                     headers={
                         'Content-Type': 'application/json',
                         'Set-Cookie': f'session={sessionCookie}',
@@ -145,7 +149,7 @@ async def validate_room_password(name: str, auth: RoomAuth):
             status_code=status.HTTP_400_BAD_REQUEST,
             detail={
                 "Code": 400,
-                "error": error
+                "Error": error
             })
 
 
@@ -153,30 +157,111 @@ async def validate_room_password(name: str, auth: RoomAuth):
 async def room(name: str, session: Optional[str] = Cookie(None)):
     try:
         if session:
-            hashed_name = parser.parse_link_hash(name)
-            room = database.get_room_by_name(hashed_name)
-            if decoder.verify_session(hashed_name, room.password, session):
-                return HTMLResponse(
-                    templates['room'],
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
+            decoded_session = await decoder.decode_session(session)
+            user = await database.get_user_by_name(decoded_session['username'])
+            if decoder.verify_session(hashed_name, room.password, session) and user.status:
+                encoded_messages, messages = database.get_all_messages(room.id), {
+                }
+                for message in encoded_messages:
+                    user = database.get_user_by_id(message.user_id)
+                    messages[user.name] = {
+                        'Message': decoder.dcrypt_message(message.data, decoded_session['msg_key']),
+                        'Created_at': parser.parse_msg_time(message.created_at)
+                    }
+                return JSONResponse(
                     status_code=status.HTTP_200_OK,
+                    content={
+                        'User': decoded_session['username'],
+                        'Messages': messages
+                    },
                     headers={
+                        'Content-Type': 'application/json',
                         'Connection': 'keep-alive'
                     }
                 )
-        return HTMLResponse(
-            templates['room_auth'],
+        return JSONResponse(
             status_code=status.HTTP_302_FOUND,
+            content={"Unauthorized": "Session doesn't exist"},
             headers={
                 'Location': f'/rooms/{name}/auth',
                 'Connection': 'close'
             }
         )
-    except Exception as error:
+    except ValueError as error:
         raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            status_code=status.HTTP_400_BAD_REQUEST,
             detail={
-                "Code": 500,
-                "error": error
+                "Code": 400,
+                "Error": error
+            })
+
+
+@app.delete('/rooms/{name}')
+async def delete_room(name: str, session: Optional[str] = Cookie(None)):
+    try:
+        if session:
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
+            user = await database.get_user_by_name(decoder.decode_session(session)['username'])
+            if decoder.verify_session(hashed_name, room.password, session, True) and user.status:
+                await database.delete_room(room.id)
+                return JSONResponse(
+                    status_code=status.HTTP_302_FOUND,
+                    content={"Status": "Room has been deleted"},
+                    headers={
+                        'Location': '/',
+                    }
+                )
+        return JSONResponse(
+            status_code=status.HTTP_302_FOUND,
+            content={"Unauthorized": "Session doesn't exist"},
+            headers={
+                'Location': f'/rooms/{name}/auth',
+                'Connection': 'close'
+            }
+        )
+    except ValueError as error:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail={
+                "Code": 400,
+                "Error": error
+            })
+
+
+@app.patch('/rooms/{name}')
+async def block_room_user(name: str, user: UserBlock, session: Optional[str] = Cookie(None)):
+    try:
+        if session:
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
+            username = jsonable_encoder(user)['username']
+            user = await database.get_user_by_name(username)
+            if decoder.verify_session(hashed_name, room.password, session, True) and user.status:
+                await database.block_user(username, room.id)
+                return JSONResponse(
+                    status_code=status.HTTP_302_FOUND,
+                    content={"Status": "User has been blocked"},
+                    headers={
+                        'Location': '/',
+                    }
+                )
+        return JSONResponse(
+            status_code=status.HTTP_302_FOUND,
+            content={"Unauthorized": "Session doesn't exist"},
+            headers={
+                'Location': f'/rooms/{name}/auth',
+                'Connection': 'close'
+            }
+        )
+    except ValueError as error:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail={
+                "Code": 400,
+                "Error": error
             })
 
 
@@ -184,37 +269,32 @@ async def room(name: str, session: Optional[str] = Cookie(None)):
 async def create_msg_key(name: str, keyData: MsgKeysCreate, session: Optional[str] = Cookie(None)):
     try:
         if session:
-            hashed_name = parser.parse_link_hash(name)
-            room = database.get_room_by_name(hashed_name)
-            encoded_data = jsonable_encoder(keyData)
-            if decoder.verify_session(hashed_name, room.password, session, True):
-                try:
-                    database.create_msg_key(
-                        room.id, encoded_data['destinied_for'], encoded_data['key']
-                    )
-                    return JSONResponse(
-                        content={"Status": "Msg created"},
-                        status_code=status.HTTP_200_OK
-                    )
-                except ValueError as error:
-                    return JSONResponse(
-                        content={"error": error},
-                        status_code=status.HTTP_400_BAD_REQUEST,
-                    )
-        return HTMLResponse(
-            templates['room_auth'],
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
+            encoded_data = await jsonable_encoder(keyData)
+            user = await database.get_user_by_name(decoder.decode_session(session)['username'])
+            if decoder.verify_session(hashed_name, room.password, session, True) and user.status:
+                database.create_msg_key(
+                    room.id, encoded_data['destinied_for'], encoded_data['key']
+                )
+                return JSONResponse(
+                    content={"Status": "Msg created"},
+                    status_code=status.HTTP_200_OK
+                )
+        return JSONResponse(
             status_code=status.HTTP_302_FOUND,
+            content={"Unauthorized": "Session doesn't exist"},
             headers={
                 'Location': f'/rooms/{name}/auth',
                 'Connection': 'close'
             }
         )
-    except Exception as error:
+    except ValueError as error:
         raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            status_code=status.HTTP_400_BAD_REQUEST,
             detail={
-                "Code": 500,
-                "error": error
+                "Code": 400,
+                "Error": error
             })
 
 
@@ -222,37 +302,41 @@ async def create_msg_key(name: str, keyData: MsgKeysCreate, session: Optional[st
 async def delete_msg_key(name: str, keyData: MsgKeysGet, session: Optional[str] = Cookie(None)):
     try:
         if session:
-            hashed_name = parser.parse_link_hash(name)
-            room = database.get_room_by_name(hashed_name)
+            hashed_name = await parser.parse_link_hash(name)
+            room = await database.get_room_by_name(hashed_name)
             encoded_data = jsonable_encoder(keyData)
-            if decoder.verify_session(hashed_name, room.password, session):
+            user = await database.get_user_by_name(decoder.decode_session(session)['username'])
+            if decoder.verify_session(hashed_name, room.password, session) and user.status:
                 try:
                     data = database.get_msg_key(
                         room.id, encoded_data['destinied_for']
                     )
+                    key_session = decoder.session_add_key(session, data.key)
                     database.delete_key(data.id)
                     return JSONResponse(
-                        content={"Key": data.key},
-                        status_code=status.HTTP_200_OK
+                        status_code=status.HTTP_200_OK,
+                        headers={
+                            'Set-Cookie': f'session={key_session}',
+                        }
                     )
                 except ValueError as error:
                     return JSONResponse(
                         content={"error": error},
                         status_code=status.HTTP_400_BAD_REQUEST,
                     )
-        return HTMLResponse(
-            templates['room_auth'],
+        return JSONResponse(
             status_code=status.HTTP_302_FOUND,
+            content={"Unauthorized": "Session doesn't exist"},
             headers={
                 'Location': f'/rooms/{name}/auth',
                 'Connection': 'close'
             }
         )
-    except Exception as error:
+    except ValueError as error:
         raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            status_code=status.HTTP_400_BAD_REQUEST,
             detail={
-                "Code": 500,
+                "Code": 400,
                 "error": error
             })
 
@@ -260,27 +344,31 @@ async def delete_msg_key(name: str, keyData: MsgKeysGet, session: Optional[str] 
 @app.websocket("/rooms/{name}/")
 async def websocket_endpoint(websocket: WebSocket, name: str, session: Optional[str] = Cookie(None)):
     if session:
-        if manager.check_room(name):
-            manager.append_room_connection(name, websocket)
-        else:
-            manager.append_room(name, Room(name))
-            manager.append_room_connection(name, websocket)
-        room = await manager.connect_room(name, websocket)
-        user = database.get_room_by_name(parser.parse_link_hash(name))
-        decoded_session = decoder.decode_session(session)
+        decoded_session = await decoder.decode_session(session)
         username = decoded_session['username']
-        try:
-            while True:
-                data = await websocket.receive_text()
-                try:
-                    if database.create_message(
-                        encoder.encrypt_message(
-                            data, str(decoded_session['msg_key']).encode()
-                        ),user.id):
-                        await room.broadcast(f"{username} says: {data}")
-                except ValueError:
-                    await room.broadcast(f"{username} havent encryption keys")
-        except WebSocketDisconnect:
-            room.disconnect(websocket)
-            await room.broadcast(f"{username} left the chat")
-            manager.close_room(name)
+        hashed_name = await parser.parse_link_hash(name)
+        room_obj = await database.get_room_by_name(hashed_name)
+        user = await database.get_user_by_name(username, room_obj.id)
+        if decoder.verify_session(hashed_name, room_obj.password, session) and user.status:
+            if manager.check_room(name):
+                manager.append_room_connection(name, websocket)
+            else:
+                manager.append_room(name, Room(name))
+                manager.append_room_connection(name, websocket)
+            room = await manager.connect_room(name, websocket)
+            try:
+                while True:
+                    data = await websocket.receive_text()
+                    try:
+                        if database.create_message(
+                                encoder.encrypt_message(
+                                    data, str(
+                                        decoded_session['msg_key']).encode()
+                                ), user.id):
+                            await room.broadcast(f"{username} says: {data}")
+                    except ValueError:
+                        await room.broadcast(f"{username} havent encryption keys")
+            except WebSocketDisconnect:
+                room.disconnect(websocket)
+                await room.broadcast(f"{username} left the chat")
+                manager.close_room(name)
